refactor: refactor common function and variables
