+++ src/fs/constants.js
@@ -1,0 +1,4 @@
+const errMsg = 'FS operation failed';
+const targetFolder = 'src/fs/files';
+
+export { errMsg, targetFolder };
