+++ src/fs/create.js
@@ -1,9 +1,9 @@
 import path from 'path';
-import { writeFile, access, constants } from 'fs/promises';
+import { writeFile } from 'fs/promises';
+import { errMsg, targetFolder } from './constants';
+import { isFileExist } from './isFileExist';
 
 const content = 'I am fresh and young';
-const errMsg = 'FS operation failed';
-const targetFolder = 'src/fs/files';
 const targetFileName = 'fresh.txt';
 const root = path.resolve();
 const targetPath = path.join(root, targetFolder, targetFileName);
@@ -15,15 +15,6 @@ const create = async () => {
     console.error(errMsg);
   } else {
     await writeFile(targetPath, content);
-  }
-};
-
-const isFileExist = async (path) => {
-  try {
-    await access(path, constants.F_OK);
-    return true;
-  } catch {
-    return false;
   }
 };
