+++ src/fs/read.js
@@ -1,10 +1,11 @@
 import path from 'path';
-import { readFile, access, constants } from 'fs/promises';
+import { readFile } from 'fs/promises';
+import { errMsg, targetFolder } from './constants';
+import { isFileExist } from './isFileExist';
 
-const errMsg = 'FS operation failed';
 const fileName = 'fileToRead.txt';
-const targetFolder = 'src/fs/files';
-const pathToFile = path.join(path.resolve(), targetFolder, fileName);
+const rootPath = path.resolve();
+const pathToFile = path.join(rootPath, targetFolder, fileName);
 
 const read = async () => {
   const isExist = await isFileExist(pathToFile);
@@ -16,15 +17,6 @@ const read = async () => {
 
   const content = await readFile(pathToFile, { encoding: 'utf8' });
   console.log(content);
-};
-
-const isFileExist = async (path) => {
-  try {
-    await access(path, constants.F_OK);
-    return true;
-  } catch {
-    return false;
-  }
 };
 
 await read();
